+++ src/app/license/list/index.tsx
@@ -1,0 +1,1 @@
+export * from './List';
